MicroLogin/requirements.txt
Flask==2.3.2
mysql-connector-python==8.0.33
Flask-CORS==3.0.10  # Adicionando Flask-CORS
